refactor(getVersion): move entrypoint of getVersion to live-common/device
