+++ apps/cli/src/commands/deviceVersion.ts
@@ -1,6 +1,6 @@
 import { from } from "rxjs";
 import { withDevice } from "@ledgerhq/live-common/hw/deviceAccess";
-import getVersion from "@ledgerhq/live-common/hw/getVersion";
+import { getVersion } from "@ledgerhq/live-common/device/use-cases/getVersionUseCase";
 import { deviceOpt } from "../scan";
 export default {
   args: [deviceOpt],
