+++ apps/web-tools/repl/commands.ts
@@ -3,7 +3,7 @@ import getDeviceName from "@ledgerhq/live-common/hw/getDeviceName";
 import getDeviceInfo from "@ledgerhq/live-common/hw/getDeviceInfo";
 import getBatteryStatus from "@ledgerhq/live-common/hw/getBatteryStatus";
 import getAddress from "@ledgerhq/live-common/hw/getAddress/index";
-import getVersion from "@ledgerhq/live-common/hw/getVersion";
+import { getVersion } from "@ledgerhq/live-common/device/use-cases/getVersion";
 import { createDeviceSocket as socket } from "@ledgerhq/live-common/socket/index";
 import getAppAndVersion from "@ledgerhq/live-common/hw/getAppAndVersion";
 import genuineCheck from "@ledgerhq/live-common/hw/genuineCheck";
