+++ libs/ledger-live-common/src/device/use-cases/getVersionUseCase.ts
@@ -1,0 +1,5 @@
+export {
+  getVersion,
+  isBootloaderVersionSupported,
+  isHardwareVersionSupported,
+} from "../../hw/getVersion";
